Refactor : sender @ManyToOne으로 변경 및 생성자 Builder 패턴 적용
- 한 명의 회원은 여러개의 알림을 생성할 수 있으니 sender와 Notification은 다대일 관계이다.
- 생성자에서 DTO를 의존하는 코드 삭제 & Builder 패턴 적용
