+++ src/main/java/teamkiim/koffeechat/domain/notification/domain/Notification.java
@@ -1,12 +1,16 @@
 package teamkiim.koffeechat.domain.notification.domain;
 
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.cglib.core.Local;
 import teamkiim.koffeechat.domain.member.domain.Member;
-import teamkiim.koffeechat.domain.notification.service.dto.request.CreateNotificationRequest;
+import teamkiim.koffeechat.domain.notification.dto.request.CreateNotificationRequest;
 
 import java.time.LocalDateTime;
+
+import static jakarta.persistence.FetchType.LAZY;
 
 @Entity
 @Getter
@@ -19,13 +23,13 @@ public class Notification {
 
     private String eventId;            //알림 이벤트 id
 
-    @ManyToOne
+    @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "receiver_id")
     private Member receiver;            //알림을 받는 회원
 
-    @OneToOne
-    @JoinColumn(name = "member_id")
-    private Member member;              // 알림 내용에 포함될 회원
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "sender_id")
+    private Member sender;              // 알림 내용에 포함될 회원
 
     private String title;               // 알림 제목
 
@@ -43,17 +47,31 @@ public class Notification {
     /**
      * 생성자
      */
-    public Notification(CreateNotificationRequest createNotificationRequest, Member receiver, String eventId) {
+    @Builder
+    public Notification(String eventId, Member receiver, Member sender, String title, String content,
+                        String url, boolean isRead, NotificationType notificationType, LocalDateTime createdTime) {
+
         this.eventId = eventId;
         this.receiver = receiver;
-        this.member = createNotificationRequest.getMember();
-        this.title = createNotificationRequest.getTitle();
-        this.content = createNotificationRequest.getContent();
-        this.url = createNotificationRequest.getUrl();
-        this.isRead = createNotificationRequest.isRead();
-        this.notificationType = createNotificationRequest.getNotificationType();
-        this.createdTime = LocalDateTime.now();
+        this.sender = sender;
+        this.title = title;
+        this.content = content;
+        this.url = url;
+        this.isRead = isRead;
+        this.notificationType = notificationType;
+        this.createdTime = createdTime;
     }
+//    public Notification(CreateNotificationRequest createNotificationRequest, Member receiver, String eventId) {
+//        this.eventId = eventId;
+//        this.receiver = receiver;
+//        this.member = createNotificationRequest.getMember();
+//        this.title = createNotificationRequest.getTitle();
+//        this.content = createNotificationRequest.getContent();
+//        this.url = createNotificationRequest.getUrl();
+//        this.isRead = createNotificationRequest.isRead();
+//        this.notificationType = createNotificationRequest.getNotificationType();
+//        this.createdTime = LocalDateTime.now();
+//    }
 
     //== 비즈니스 로직 ==//
