+++ src/main/java/teamkiim/koffeechat/domain/notification/repository/NotificationRepository.java
@@ -8,7 +8,6 @@ import teamkiim.koffeechat.domain.notification.domain.Notification;
 
 import java.util.Optional;
 
-
 /**
  * 알림 메시지 저장 repository
  */
